requirements.txt
pycryptodome>=3.15.0
pillow>=9.0.0 